+++ lib/elastic_record/index/bulk_documents.rb
@@ -1,0 +1,86 @@
+require 'active_support/core_ext/object/to_query'
+
+module ElasticRecord
+  class Index
+    module BulkDocuments
+      def index_document(id, document, doctype: model.doctype, parent: nil, index_name: alias_name)
+        p "ARE WE RUNNING bulk???"
+        if batch = current_bulk_batch
+          instructions = { _index: index_name, _type: doctype.name, _id: id }
+          instructions[:parent] = parent if parent
+
+          batch << { index: instructions }
+          batch << document
+        else
+          super
+        end
+      end
+
+      def update_document(id, document, doctype: model.doctype, parent: nil, index_name: alias_name)
+        raise "Cannot update a document with empty id" if id.blank?
+        params = {doc: document, doc_as_upsert: true}
+
+        if batch = current_bulk_batch
+          instructions = { _index: index_name, _type: doctype.name, _id: id, _retry_on_conflict: 3 }
+          instructions[:parent] = parent if parent
+
+          batch << { update: instructions }
+          batch << params
+        else
+          super
+        end
+      end
+
+      def delete_document(id, doctype: model.doctype, parent: nil, index_name: alias_name)
+        raise "Cannot delete document with empty id" if id.blank?
+        index_name ||= alias_name
+
+        if batch = current_bulk_batch
+          instructions = { _index: index_name, _type: doctype.name, _id: id, _retry_on_conflict: 3 }
+          instructions[:parent] = parent if parent
+          batch << { delete: instructions }
+        else
+          super
+        end
+      end
+
+      def bulk(options = {})
+        connection.bulk_stack.push []
+
+        yield
+
+        if current_bulk_batch.any?
+          body = current_bulk_batch.map { |action| "#{JSON.generate(action)}\n" }.join
+          results = connection.json_post("/_bulk?#{options.to_query}", body)
+          verify_bulk_results(results)
+        end
+      ensure
+        connection.bulk_stack.pop
+      end
+
+      def bulk_add(batch, index_name: alias_name)
+        bulk do
+          batch.each do |record|
+            index_record(record, index_name: index_name)
+          end
+        end
+      end
+
+      def current_bulk_batch
+        connection.bulk_stack.last
+      end
+
+      private
+
+        def verify_bulk_results(results)
+          return unless results.is_a?(Hash)
+
+          errors = results['items'].select do |item|
+            item.values.first['error']
+          end
+
+          raise ElasticRecord::BulkError.new(errors) unless errors.empty?
+        end
+    end
+  end
+end
