+++ lib/elastic_record/index/documents.rb
@@ -75,21 +75,13 @@ module ElasticRecord
       end
 
       def index_document(id, document, doctype: model.doctype, parent: nil, index_name: alias_name)
-        if batch = current_bulk_batch
-          instructions = { _index: index_name, _type: doctype.name, _id: id }
-          instructions[:parent] = parent if parent
+        path = "/#{index_name}/#{doctype.name}/#{id}"
+        path << "?parent=#{parent}" if parent
 
-          batch << { index: instructions }
-          batch << document
+        if id
+          connection.json_put path, document
         else
-          path = "/#{index_name}/#{doctype.name}/#{id}"
-          path << "?parent=#{parent}" if parent
-
-          if id
-            connection.json_put path, document
-          else
-            connection.json_post path, document
-          end
+          connection.json_post path, document
         end
       end
 
@@ -97,34 +89,19 @@ module ElasticRecord
         raise "Cannot update a document with empty id" if id.blank?
         params = {doc: document, doc_as_upsert: true}
 
-        if batch = current_bulk_batch
-          instructions = { _index: index_name, _type: doctype.name, _id: id, _retry_on_conflict: 3 }
-          instructions[:parent] = parent if parent
-
-          batch << { update: instructions }
-          batch << params
-        else
-          path = "/#{index_name}/#{doctype.name}/#{id}/_update?retry_on_conflict=3"
-          path << "&parent=#{parent}" if parent
-
-          connection.json_post path, params
-        end
+        path = "/#{index_name}/#{doctype.name}/#{id}/_update?retry_on_conflict=3"
+        path << "&parent=#{parent}" if parent
+        connection.json_post path, params
       end
 
       def delete_document(id, doctype: model.doctype, parent: nil, index_name: alias_name)
         raise "Cannot delete document with empty id" if id.blank?
         index_name ||= alias_name
 
-        if batch = current_bulk_batch
-          instructions = { _index: index_name, _type: doctype.name, _id: id, _retry_on_conflict: 3 }
-          instructions[:parent] = parent if parent
-          batch << { delete: instructions }
-        else
-          path = "/#{index_name}/#{doctype.name}/#{id}"
-          path << "&parent=#{parent}" if parent
+        path = "/#{index_name}/#{doctype.name}/#{id}"
+        path << "&parent=#{parent}" if parent
 
-          connection.json_delete path
-        end
+        connection.json_delete path
       end
 
       def delete_by_query(query)
@@ -163,43 +140,6 @@ module ElasticRecord
         connection.json_get("/_search/scroll?#{options.to_query}")
       end
 
-      def bulk(options = {})
-        connection.bulk_stack.push []
-
-        yield
-
-        if current_bulk_batch.any?
-          body = current_bulk_batch.map { |action| "#{JSON.generate(action)}\n" }.join
-          results = connection.json_post("/_bulk?#{options.to_query}", body)
-          verify_bulk_results(results)
-        end
-      ensure
-        connection.bulk_stack.pop
-      end
-
-      def bulk_add(batch, index_name: alias_name)
-        bulk do
-          batch.each do |record|
-            index_record(record, index_name: index_name)
-          end
-        end
-      end
-
-      def current_bulk_batch
-        connection.bulk_stack.last
-      end
-
-      private
-
-        def verify_bulk_results(results)
-          return unless results.is_a?(Hash)
-
-          errors = results['items'].select do |item|
-            item.values.first['error']
-          end
-
-          raise ElasticRecord::BulkError.new(errors) unless errors.empty?
-        end
     end
   end
 end
